Move DeepSeekV3 skeleton to models/deepseek_v3 subpackage
