+++ transformer_engine/pytorch/models/deepseek_v3/__init__.py
@@ -4,8 +4,10 @@
 
 """DeepSeekV3 transformer layer built from Transformer Engine MoE building blocks."""
 
-from transformer_engine.pytorch.deepseek.multi_latent_attention import MultiLatentAttention
-from transformer_engine.pytorch.deepseek.moe import DeepSeekV3MoE
-from transformer_engine.pytorch.deepseek.transformer_layer import DeepSeekV3Layer
+from transformer_engine.pytorch.models.deepseek_v3.multi_latent_attention import (
+    MultiLatentAttention,
+)
+from transformer_engine.pytorch.models.deepseek_v3.moe import DeepSeekV3MoE
+from transformer_engine.pytorch.models.deepseek_v3.transformer_layer import DeepSeekV3Layer
 
 __all__ = ["DeepSeekV3Layer", "DeepSeekV3MoE", "MultiLatentAttention"]
